add engine testing for both INTERNAL and FILEIO
